moved the validator to its own file
